Revert to using BigInteger for certificate serials.
We'll still use a long db sequence to generate them, but storing as big
integer prevents having to do unsafe casts in other areas.

+++ proxy/src/main/java/org/fedoraproject/candlepin/model/EntitlementCertificate.java
@@ -14,6 +14,8 @@
  */
 package org.fedoraproject.candlepin.model;
 
+import java.math.BigInteger;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -54,18 +56,18 @@ public class EntitlementCertificate implements Persisted {
     private byte[] cert;
 
     @Column(nullable = false)
-    private Long serial;
+    private BigInteger serial;
 
     @ManyToOne
     @ForeignKey(name = "fk_cert_entitlement")
     @JoinColumn(nullable = false)
     private Entitlement entitlement;
 
-    public Long getSerial() {
+    public BigInteger getSerial() {
         return serial;
     }
 
-    public void setSerial(Long serialNumber) {
+    public void setSerial(BigInteger serialNumber) {
         this.serial = serialNumber;
     }
