+++ proxy/src/test/java/org/fedoraproject/candlepin/test/DatabaseTestFixture.java
@@ -14,6 +14,7 @@
  */
 package org.fedoraproject.candlepin.test;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 import javax.persistence.EntityManager;
@@ -207,7 +208,7 @@ public class DatabaseTestFixture {
     }
     
     protected EntitlementCertificate createEntitlementCertificate(String key, String cert, 
-            Long serial) {
+            BigInteger serial) {
         EntitlementCertificate toReturn = new EntitlementCertificate();
         toReturn.setKey(key.getBytes());
         toReturn.setCert(cert.getBytes());
